+++ python_password/PyPassword.py
@@ -21,12 +21,13 @@ from python_password.utils.crypto import *
 from python_password.utils.database import *
 from python_password.utils.files import *
 from python_password.utils.settings import *
+from translations.core import *
 
 
 class SimpleDialog:
     """Simple dialogs."""
 
-    def __init__(self, title, text, alert_text='OK', **kwargs):
+    def __init__(self, title, text, alert_text, **kwargs):
         super().__init__(**kwargs)
         self.title = title
         self.text = text
@@ -71,23 +72,30 @@ class PyPassword(MDApp):
         super().__init__(**kwargs)
         self.theme_cls.primary_palette = 'Indigo'
         self.theme_cls.theme_style = 'Light'
-        self.text_color_hex = '111111'
+        self.text_color_hex = get_setting('text_color_hex')
         self.passwords = []
         self.info = {
             'name': 'Python Password',
-            'version': '0.2.4',
+            'version': '0.2.5',
             'author': 'Jakub Suchenek',
             'github': 'https://github.com/AnonymousX86/Python-Password',
             'faq': 'https://github.com/AnonymousX86/Python-Password/blob/master/docs/FAQ.md',
             'mail': 'mailto:jakub.suchenek.25@gmail.com',
             'icon': 'Icon made by Freepik from www.flaticon.com',
-            'rd_party': 'UPX, Kivy and KivyMD'
+            '3rd_party': 'UPX, Kivy and KivyMD'
         }
+        self.all_languages = {
+            'en': 'English',
+            'pl': 'Polski'
+        }
+        self.current_language = get_setting('language')
+        self.messages = {}
 
     # It has to be outside ``__init__``
-    text_color_rgba = ListProperty([.06, .06, .06, 1])
+    text_color_rgba = ListProperty(get_setting('text_color_rgba'))
 
     def build(self):
+        self.update_messages()
         with open(f'kv_templates{sep}PyPassword.kv', encoding='utf8') as fd:
             kv = Builder.load_string(fd.read())
         return kv
@@ -96,6 +104,9 @@ class PyPassword(MDApp):
         self.update_passwords_list()
         self.masters_ok()
         self.switch_theme(get_theme())
+        self.update_languages_list()
+        # Development option
+        # self.root.ids.screen_manager.current = 'languages'
 
     # ================================
     #           Information
@@ -105,12 +116,12 @@ class PyPassword(MDApp):
         """Shows dialog with options what to do with password."""
         ctx_dialog = MDDialog(
             title=f'[color=#{self.text_color_hex}]{instance.text.capitalize()}[/color]',
-            text='Choose what do you want to do wth this password.',
+            text=self.tr('ctx_text'),
             auto_dismiss=False,
             buttons=[
                 MDRectangleFlatIconButton(
                     icon='content-copy',
-                    text='Copy',
+                    text=self.tr('copy'),
                     on_release=lambda x: [
                         self.copy_password(instance.text),
                         ctx_dialog.dismiss()
@@ -118,7 +129,7 @@ class PyPassword(MDApp):
                 ),
                 MDRectangleFlatIconButton(
                     icon='trash-can-outline',
-                    text='Delete',
+                    text=self.tr('delete'),
                     on_release=lambda x: [
                         self.del_password(instance.text, True),
                         ctx_dialog.dismiss()
@@ -126,7 +137,7 @@ class PyPassword(MDApp):
                 ),
                 MDRectangleFlatIconButton(
                     icon='arrow-left-circle-outline',
-                    text='Nothing',
+                    text=self.tr('nothing'),
                     on_release=lambda x: ctx_dialog.dismiss()
                 )
             ]
@@ -136,8 +147,7 @@ class PyPassword(MDApp):
     def detailed_info(self, name: str):
         """Opens dialog about specific info about program in ``info`` screen."""
         info_dialog = MDDialog(
-            title=f'[color={self.text_color_hex}]' +
-                  (name.capitalize() if name != 'rd_party' else '3rd party software') + '[/color]',
+            title=self.tr(name, is_title=True),
             text=self.info[name],
             auto_dismiss=False,
             buttons=[
@@ -154,8 +164,9 @@ class PyPassword(MDApp):
 
     def wip_info(self):
         info_dialog = SimpleDialog(
-            title='Work in progress',
-            text='This feature is under development.'
+            title=self.tr('wip_title'),
+            text=self.tr('wip_text'),
+            alert_text=self.tr('ok')
         ).alert()
         info_dialog.open()
 
@@ -171,21 +182,24 @@ class PyPassword(MDApp):
             ok_value = self.validate_input(value_box, 6)
             if ok_alias is ValueTooShort or ok_value is ValueTooShort:
                 result_dialog = SimpleDialog(
-                    title='Whoops!',
-                    text='At least one value is too short.'
+                    title=self.tr('whoops'),
+                    text=self.tr('error_input_too_short'),
+                    alert_text=self.tr('ok')
                 ).alert()
             elif ok_alias is PatternError or ok_value is PatternError:
                 result_dialog = SimpleDialog(
-                    title='Whoops!',
-                    text='At last one value is invalid.'
+                    title=self.tr('whoops'),
+                    text=self.tr('error_input_invalid'),
+                    alert_text=self.tr('ok')
                 ).alert()
             else:
                 password_alias = alias_box.text.capitalize()
                 password_value = value_box.text
                 if already_exists(password_alias):
                     result_dialog = SimpleDialog(
-                        title='Whoops!',
-                        text='That password already exists.'
+                        title=self.tr('whoops'),
+                        text=self.tr('error_already_exists'),
+                        alert_text=self.tr('ok')
                     ).alert()
                 else:
                     save_password(
@@ -193,8 +207,9 @@ class PyPassword(MDApp):
                         encrypt(password_value)
                     )
                     result_dialog = SimpleDialog(
-                        title='Success!',
-                        text=f'Password "{password_alias}" successfully saved.'
+                        title=self.tr('success'),
+                        text=self.tr('success_save', txt_format=password_alias),
+                        alert_text=self.tr('ok')
                     ).alert()
             result_dialog.open()
         alias_box.text = ''
@@ -205,32 +220,36 @@ class PyPassword(MDApp):
         to_decrypt = get_one_password(password)
         if to_decrypt is None:
             result_dialog = SimpleDialog(
-                title='Warning!',
-                text='That password do not exists in database.'
+                title=self.tr('warning'),
+                text=self.tr('error_not_exists'),
+                alert_text=self.tr('ok')
             ).alert()
         elif type(to_decrypt) is not bytes:
             result_dialog = SimpleDialog(
-                title='Warning!',
-                text='An critical error has occurred. Passwords are saved to local database in wrong way.'
+                title=self.tr('warning'),
+                text=self.tr('error_bad_format'),
+                alert_text=self.tr('ok')
             ).alert()
         else:
             decrypted = decrypt(to_decrypt)
             if type(decrypted) is InvalidToken:
                 result_dialog = SimpleDialog(
-                    title='Warning!',
-                    text='Alpha password do not match. If you want to access this password,'
-                         ' please change alpha password. If needed, change beta password too.'
+                    title=self.tr('warning'),
+                    text=self.tr('error_invalid_token'),
+                    alert_text=self.tr('ok')
                 ).alert()
             elif type(decrypted) is not str:
                 result_dialog = SimpleDialog(
-                    title='Warning!',
-                    text=f'An error has occurred: {decrypted}'
+                    title=self.tr('warning'),
+                    text=self.tr('error_unknown', txt_format=decrypted),
+                    alert_text=self.tr('ok')
                 ).alert()
             else:
                 copy(decrypted)
                 result_dialog = SimpleDialog(
-                    title='Success!',
-                    text='Password copied to clipboard. Now you can paste in somewhere.'
+                    title=self.tr('success'),
+                    text=self.tr('success_copy'),
+                    alert_text=self.tr('ok')
                 ).alert()
         result_dialog.open()
         self.update_passwords_list()
@@ -241,34 +260,38 @@ class PyPassword(MDApp):
         self.root.ids.del_password_alias.text = ''
         if not force and len(password) == 0:
             result_dialog = SimpleDialog(
-                title='Whoops!',
-                text='Please provide password alias at first.'
+                title=self.tr('whoops'),
+                text=self.tr('error_no_alias'),
+                alert_text=self.tr('ok')
             ).alert()
         elif not force and len(password) < 3:
             result_dialog = SimpleDialog(
-                title='Whoops!',
-                text='Password alias has to be at least 3 characters long.'
+                title=self.tr('whoops'),
+                text=self.tr('error_alias_too_short', txt_format='3'),
+                alert_text=self.tr('ok')
             ).alert()
         else:
             to_del = get_one_password(password)
             if to_del is None:
                 result_dialog = SimpleDialog(
-                    title='Warning!',
-                    text='That password do not exists in database'
+                    title=self.tr('warning'),
+                    text=self.tr('error_not_exists'),
+                    alert_text=self.tr('ok')
                 ).alert()
             elif type(to_del) is not bytes:
                 result_dialog = SimpleDialog(
-                    title='Warning!',
-                    text='An critical error has occurred. Passwords are saved to local database in wrong way.'
+                    title=self.tr('warning'),
+                    text=self.tr('error_bad_format'),
+                    alert_text=self.tr('ok')
                 ).alert()
             else:
                 result_dialog = MDDialog(
-                    title='Attention',
-                    text=f'Do you really want to delete "{password}" password?',
+                    title=self.tr('warning'),
+                    text=self.tr('confirm_delete', txt_format=password),
                     auto_dismiss=False,
                     buttons=[
                         MDFillRoundFlatIconButton(
-                            text='Yes',
+                            text=self.tr('yes'),
                             icon='check-circle-outline',
                             on_release=lambda x: [
                                 self._del_password_confirm(password),
@@ -276,7 +299,7 @@ class PyPassword(MDApp):
                             ]
                         ),
                         MDRoundFlatIconButton(
-                            text='No',
+                            text=self.tr('no'),
                             icon='close-circle-outline',
                             on_release=lambda x: self.dismiss_and_back(result_dialog)
                         )
@@ -287,8 +310,9 @@ class PyPassword(MDApp):
     def _del_password_confirm(self, password: str):
         del_password(password)
         result_dialog = SimpleDialog(
-            title='Success!',
-            text=f'Password "{password}" successfully deleted.'
+            title=self.tr('success'),
+            text=self.tr('success_delete', txt_format=password),
+            alert_text=self.tr('ok')
         ).alert()
         result_dialog.open()
         self.update_passwords_list()
@@ -301,18 +325,17 @@ class PyPassword(MDApp):
         """Checks if master passwords are OK."""
         if not check_beta():
             alert_dialog = MDDialog(
-                title='Missing beta password',
-                text='I\'ve noticed, that you have not set beta password. It\'s needed for safe password storing.'
-                     ' Do you want to provide it by yourself, or let program to randomize it?',
+                title=self.tr('error_missing_master_title', txt_format=self.tr('beta')),
+                text=self.tr('error_missing_master_text'),
                 auto_dismiss=False,
                 buttons=[
                     MDFillRoundFlatIconButton(
-                        text='Set password',
+                        text=self.tr('set_password'),
                         icon='account-key-outline',
                         on_release=lambda x: self.dismiss_and_back(alert_dialog, 'settings')
                     ),
                     MDRoundFlatIconButton(
-                        text='Randomize',
+                        text=self.tr('randomize'),
                         icon='dice-multiple-outline',
                         on_release=lambda x: [
                             self.reset_beta(),
@@ -325,18 +348,17 @@ class PyPassword(MDApp):
             return False
         elif not check_alpha():
             alert_dialog = MDDialog(
-                title='Missing alpha password',
-                text='I\'ve noticed, that you have not set alpha password. It\'s needed for safe password storing.'
-                     ' Do you want to provide it by yourself, or let program to randomize it?',
+                title=self.tr('error_missing_master_title', txt_format=self.tr('alpha')),
+                text=self.tr('error_missing_master_text'),
                 auto_dismiss=False,
                 buttons=[
                     MDFillRoundFlatIconButton(
-                        text='Set password',
+                        text=self.tr('set_password'),
                         icon='account-key-outline',
                         on_release=lambda x: self.dismiss_and_back(alert_dialog, 'settings')
                     ),
                     MDRoundFlatIconButton(
-                        text='Randomize',
+                        text=self.tr('randomize'),
                         icon='dice-multiple-outline',
                         on_release=lambda x: [
                             self.reset_alpha(),
@@ -366,19 +388,20 @@ class PyPassword(MDApp):
         if len(password) < 6:
             password_box.error = True
             result_dialog = SimpleDialog(
-                title='Whoops!',
-                text=f'{which.capitalize()} password should be at least 6 characters long.'
+                title=self.tr('whoops'),
+                text=self.tr('error_master_too_short', txt_format=[self.tr(which), '6']),
+                alert_text=self.tr('ok')
             ).alert()
         else:
             password_box.error = False
             generate_func(password)
             result_dialog = MDDialog(
-                title='Success!',
-                text=f'New {which} password successfully saved.',
+                title=self.tr('success'),
+                text=self.tr('success_new_master', txt_format=self.tr(which)),
                 auto_dismiss=False,
                 buttons=[
                     MDRaisedButton(
-                        text='OK',
+                        text=self.tr('ok'),
                         on_release=lambda x: self.dismiss_and_back(result_dialog)
                     )
                 ]
@@ -389,23 +412,21 @@ class PyPassword(MDApp):
 
     def reset_alpha(self):
         confirm_dialog = MDDialog(
-            title='Warning!',
-            text='This will randomize alpha password. It will be super safe, but you probably will not be able to'
-                 ' re-enter this password later. You will not be able to copy already saved passwords.'
-                 ' Do you want to continue?',
+            title=self.tr('warning'),
+            text=self.tr('confirm_randomize_alpha'),
             auto_dismiss=False,
             buttons=[
                 MDFillRoundFlatIconButton(
+                    text=self.tr('yes'),
                     icon='check-circle-outline',
-                    text='Yes',
                     on_release=lambda x: [
                         self.change_master('alpha', urandom(16)),
                         confirm_dialog.dismiss()
                     ]
                 ),
                 MDRoundFlatIconButton(
+                    text=self.tr('no'),
                     icon='close-circle-outline',
-                    text='No',
                     on_release=lambda x: confirm_dialog.dismiss()
                 )
             ]
@@ -414,22 +435,21 @@ class PyPassword(MDApp):
 
     def reset_beta(self):
         confirm_dialog = MDDialog(
-            title='Warning!',
-            text='This will randomize beta password. It will be super safe, but you probably will not be able to'
-                 ' re-enter this password later. Do you want to continue?',
+            title=self.tr('warning'),
+            text=self.tr('confirm_randomize_beta'),
             auto_dismiss=False,
             buttons=[
                 MDFillRoundFlatIconButton(
+                    text=self.tr('yes'),
                     icon='check-circle-outline',
-                    text='Yes',
                     on_release=lambda x: [
                         self.change_master('beta', urandom(16)),
                         confirm_dialog.dismiss()
                     ]
                 ),
                 MDRoundFlatIconButton(
+                    text=self.tr('no'),
                     icon='close-circle-outline',
-                    text='No',
                     on_release=lambda x: confirm_dialog.dismiss()
                 )
             ]
@@ -470,27 +490,27 @@ class PyPassword(MDApp):
     def _set_passwords_count(self):
         count = len(self.passwords)
         if count == 0:
-            text = 'There are no passwords in database.'
+            text = self.tr('password_count_0')
             self.root.ids.passwords_list.add_widget(
                 NewbieTip(
-                    text='Welcome to Python Password',
-                    secondary_text='Save your first password, by entering',
-                    tertiary_text='it\'s data on the right',
+                    text=self.tr('tip_1_title', txt_format=self.info['name']),
+                    secondary_text=self.tr('tip_1_text_1'),
+                    tertiary_text=self.tr('tip_1_text_2'),
                     icon='arrow-right-bold-circle'
                 )
             )
         elif count == 1:
-            text = 'There\'s only 1 password in database.'
+            text = self.tr('password_count_1')
             self.root.ids.passwords_list.add_widget(
                 NewbieTip(
-                    text='Congratulations!',
-                    secondary_text='You have saved your\'s 1st password.',
-                    tertiary_text='Preview it just by clicking it.',
+                    text=self.tr('tip_2_title'),
+                    secondary_text=self.tr('tip_2_text_1'),
+                    tertiary_text=self.tr('tip_2_text_2'),
                     icon='arrow-up-bold-circle'
                 )
             )
         else:
-            text = f'There are {count} passwords in database.'
+            text = self.tr('passwords_count_x', txt_format=count)
         self.root.ids.passwords_count.text = text
 
     def update_passwords_list(self):
@@ -499,8 +519,81 @@ class PyPassword(MDApp):
         self._set_passwords_count()
 
     # ================================
+    #           Languages
+    # ================================
+
+    def _set_languages_list(self):
+        self.root.ids.languages_list.clear_widgets()
+        for lang in self.all_languages.keys():
+            self.root.ids.languages_list.add_widget(
+                OneLineListItem(
+                    text=self.all_languages[lang],
+                    on_release=lambda x: self.change_language(
+                        list(self.all_languages.keys())[list(self.all_languages.values()).index(x.text)]
+                    )
+                )
+            )
+
+    def update_languages_list(self):
+        self._set_languages_list()
+
+    def update_messages(self):
+        self.messages = get_messages(get_setting('language'))
+
+    def change_language(self, lang: str):
+        if not check_language(lang):
+            result_dialog = MDDialog(
+                title=self.tr('whoops'),
+                text=self.tr('error_locale_missing'),
+                auto_dismiss=False,
+                buttons=[
+                    MDRaisedButton(
+                        text=self.tr('yes'),
+                        on_release=lambda x: [
+                            download_locale(lang),
+                            set_locale(lang),
+                            self.restart_required(),
+                            result_dialog.dismiss()
+                        ]
+                    ),
+                    MDRaisedButton(
+                        text=self.tr('no'),
+                        on_release=lambda x: result_dialog.dismiss()
+                    )
+                ]
+            )
+            result_dialog.open()
+        else:
+            set_locale(lang)
+            self.restart_required()
+
+    def add_language(self):
+        self.open_url(
+            'https://github.com/AnonymousX86/Python-Password/blob/master/docs/CONTRIBUTING.md#adding-new-locales'
+        )
+
+    def tr(self, text_id, txt_format=None, is_title=False):
+        try:
+            result = self.messages[text_id]
+        except KeyError:
+            try:
+                result = default_messages[text_id]
+            except KeyError:
+                return f'Missing: {text_id}'
+        if txt_format:
+            result = result.format(txt_format)
+        if is_title:
+            result = f'[color=#{get_setting("text_color_hex")}]{result}[/color]'
+        return result
+
+    # ================================
     #               Misc
     # ================================
+
+    def update_theme(self):
+        self.text_color_hex = get_setting('text_color_hex')
+        self.text_color_rgba = get_setting('text_color_rgba')
+        self.theme_cls.theme_style = get_setting('theme')
 
     def switch_theme(self, force=None):
         """
@@ -517,16 +610,22 @@ class PyPassword(MDApp):
 
         if current == 'Light':
             self.theme_cls.theme_style = 'Dark'
-            self.text_color_hex = 'ffffff'
-            self.text_color_rgba = (1, 1, 1, 1)
             set_theme('Dark')
         elif current == 'Dark':
             self.theme_cls.theme_style = 'Light'
-            self.text_color_hex = '111111'
-            self.text_color_rgba = (.06, .06, .06, 1)
             set_theme('Light')
         else:
             raise NameError('No theme found')
+
+        self.update_theme()
+
+    def restart_required(self):
+        result_dialog = SimpleDialog(
+            title=self.tr('success'),
+            text=self.tr('restart_required'),
+            alert_text=self.tr('ok')
+        ).alert()
+        result_dialog.open()
 
     def open_url(self, url: str):
         """Opens URL in default browser."""
@@ -545,7 +644,10 @@ class PyPassword(MDApp):
         if len(instance.text) < length:
             instance.error = True
             return ValueTooShort
-        elif match('^[A-Za-z0-9][A-Za-z0-9 &\\-_]+[A-Za-z0-9]$', instance.text) is None:
+        elif match(
+                '^[A-Za-zĘÓĄŚŁŻŹĆŃęóąśłżźćń0-9][A-Za-z0-9ĘÓĄŚŁŻŹĆŃęóąśłżźćń &\\-_]+[A-Za-z0-9ĘÓĄŚŁŻŹĆŃęóąśłżźćń]$',
+                instance.text
+        ) is None:
             instance.error = True
             return PatternError
         else: